Rename pkg/server to pkg/api
