+++ pkg/api/grpcapi/internal/pb/service_admin.proto
@@ -1,7 +1,7 @@
 syntax = "proto3";
 package pb;
 
-import "pkg/server/grpcapi/internal/pb/user.proto";
+import "pkg/api/grpcapi/internal/pb/user.proto";
 
 service Admin {
